Lithium-Ion Battery management system
A battery management system for lithium-ion cells. It is used to keep battery operation more efficient and safe. This circuit is for rechargeable batteries. It is going to monitor the state of the battery, calculating the secondary data and controlling the cells and balances it. In this circuit, it monitors the various state of the battery such as voltage, temperature, state of charge and depth of discharge. 
This battery management system will protect the battery from over-discharge, overcharge, and short circuit protection.