Student attendance monitoring system
A student attendance monitoring system can be used in every organization to mark the attendance of students or employees. The main application of the Student attendance monitoring system is seen in teaching institutions, where the attendance of students has to be regularly monitored on an hourly basis. The method developed provides a secure and effective may recording attendance. Student attendance monitoring system uses mainly two algorithms i.e. Viola-Jones Algorithm and Local binary pattern algorithm. Viola-Jones algorithm is used for face detection and Local binary pattern is used for the feature extraction and face recognition. For classification support, a vector machine algorithm is used.