Efficient Energy management system using IoT
In this project work, “Efficient Energy Management system using IoT” the module collects energy consumption information from each device and further processes and analyzes it. Since lighting and air conditioning appliances contribute to most of the electricity consumption in the buildings, monitoring them is more important. The sun position at a given time and place is calculated in the cloud and the result will be sent to the MCU. From the received signal the MCU will send the control signal to the solar positioning system. The MCU will control the charging of the power wall from the solar panel. The MCU will monitor the current and voltage of the battery. The condition of the energy stored in the battery will be set back to the cloud. The user will be having a web application to view the conditions of the power wall and control the appliances.